region cleaning refactoring
